Change every to one file
